+++ Code/Scar/Scar/UIObject.cpp
@@ -3,113 +3,6 @@
 
 
 //////////////////////////////////////////////////////////////////
-//名称：UIAnima
-//描述：实现UI的平移，旋转，缩放，Alpah改变等动画
-//作者：屠文翔，杨成熙
-//////////////////////////////////////////////////////////////////
-//
-////构造函数
-//UIAnima::UIAnima( UIObject* host, u32 duration, u32 loop, ANIMA_END_OPTION opflag, u32 interval )
-//{
-//	Host = host;
-//	SetUIAnima( duration, loop, opflag, interval);
-//	LastFrameTime = MyIrrlichtEngine::GetEngine()->GetDevice()->getTimer()->getRealTime();
-//	CurrentFrame = 1;
-//	CurrentLoop = 1;
-//	HasRotate = HasScale = HasTranslate = false;
-//}
-////设置动画信息
-//void UIAnima::SetUIAnima( u32 duration, u32 loop, ANIMA_END_OPTION opflag, u32 interval)
-//{
-//	Duration = duration;
-//	Loop = loop;
-//	Interval = interval;
-//	EndOption = opflag;
-//	NumOfFrame = Duration / Interval;
-//}
-////添加缩放动画
-//void UIAnima::AddScale( const vector2d<f32>& stretch, const vector2d<f32>& scalePoint )
-//{
-//	HasScale = true;
-//	StepStr.X = stretch.X / NumOfFrame;
-//	StepStr.Y = stretch.Y / NumOfFrame;
-//	StrCen = scalePoint;
-//}
-////添加旋转动画
-//void UIAnima::AddRotate( float degree, const vector2d<f32>& rotatePoint )
-//{
-//	HasRotate = true;
-//	StepDeg = degree / NumOfFrame;
-//	RotCen = rotatePoint;
-//}
-////添加平移动画
-//void AddTranslate( vector2d<f32> offset );
-////步进缩放
-//void Scale( vector2d<f32> stepstr, vector2d<f32> scalePoint );
-////步进旋转
-//void UIAnima::Rotate( float stepdeg, const vector2d<f32>& rotatePoint )
-//{
-//	float steprad = stepdeg / 180 * PI;
-//	//正角度为逆时针
-//	//记录矩阵四个顶点临时坐标
-//	vector2d<f32> temQuar[4];
-//	//得到旋转后的坐标
-//	for (int i = 0; i < 4; i++)
-//	{
-//		temQuar[i] = Host->DstQuar[i] - rotatePoint;
-//		float x,y;
-//		x = (float)temQuar[i].X;
-//		y = (float)temQuar[i].Y;
-//		temQuar[i].X = cos(steprad) * x - sin(steprad) * y;
-//		temQuar[i].Y = sin(steprad) * x + cos(steprad) * y;
-//		//加回到原来的坐标系
-//		Host->DstQuar[i] = temQuar[i] + rotatePoint;
-//	}
-//}
-////步进平移
-//void Translate( vector2d<f32> steptran );
-////运行动画
-//void UIAnima::Run()
-//{
-//	//控制帧频
-//	u32 now = MyIrrlichtEngine::GetEngine()->GetDevice()->getTimer()->getRealTime();
-//	if(now - LastFrameTime < Interval )
-//		return;
-//	LastFrameTime = now;
-//
-//	//应用变换
-//	if (HasScale)
-//	{
-//	}
-//	if (HasRotate)
-//	{
-//		Rotate( StepDeg, RotCen );
-//	}
-//
-//	//控制帧数
-//	if(HasRotate || HasScale || HasTranslate)
-//	{
-//		CurrentFrame++;
-//		if(CurrentFrame > NumOfFrame)
-//		{
-//			if(Loop == 0)
-//			{
-//				CurrentFrame = 1;
-//			}
-//			else
-//			{
-//				CurrentLoop++;
-//				if(CurrentLoop > Loop)
-//				{
-//					if(EndOption == STAY)
-//						HasTranslate = HasScale = HasRotate = false;
-//				}
-//			}
-//		}
-//	}
-//}
-
-//////////////////////////////////////////////////////////////////
 //名称：UIObject
 //描述：UI基类
 //作者：屠文翔，杨成熙
@@ -153,12 +46,12 @@ const vector2d<f32>& UIObject::GetCenter() const
 	return Center;
 }
 
-unsigned int UIObject::GetAlpha()
+f32 UIObject::GetAlpha()
 {
 	return Alpha;
 }
 
-void UIObject::SetAlpha( unsigned int alpha )
+void UIObject::SetAlpha( f32 alpha )
 {
 	Alpha = alpha;
 }
@@ -184,7 +77,7 @@ void UIImage::Draw()
 	SColor colors[4];
 	for (int i=0; i<4; i++)
 	{
-		colors[i] = SColor(Alpha,255,255,255);
+		colors[i] = SColor((u32)Alpha,255,255,255);
 	}
 	Driver->draw2DImage( Image, intDstQuar, rect<s32>(0,0,w,h), 0, colors/*&SColor(Alpha,255,255,255)*/, true );
 	//Driver->draw2DImage( Image, vector2d<s32>(0,0));
